+++ Runtime/AssemblyInfo.cs
@@ -1,0 +1,5 @@
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("Creator-Hian.Unity.Singleton.Tests")]
+[assembly: InternalsVisibleTo("Creator-Hian.Unity.Singleton.Tests.Runtime")]
+[assembly: InternalsVisibleTo("Creator-Hian.Unity.Singleton.Tests.Editor")]